.ipynb_checkpoints/requirements-checkpoint.txt
numpy
torch
torchvision
diffusers==0.30.2
ultralytics
Pillow
uuid
accelerate
safetensors
opencv-python
transformers
sentencepiece
protobuf
peft

# You can also add Git repos as dependencies, but you'll need to add git to the system_packages list in cog.yaml:
# 
# build:
#   system_packages:
#     - "git"
# 
# Then you can use a URL like this:
# 
# git+https://github.com/huggingface/transformers


# You can also pin Git repos to a specific commit:
# 
# git+https://github.com/huggingface/transformers@2d1602a
